Remove dependency on a running play application.
Play's WS client depends on a running Play application, which coupled
this jsonhomeclient directly to Play.

The JsonHomeClient now expects a WSClient, in a Play app this can be
created via `WS.client`, in a non-Play app it can be created via
`new NingWSClient(new AsyncHttpClientConfig.Builder().build())` (remember
to `close()` the client in this case).

Because there's no longer a running Play app required, the JsonHomeCacheSpec
could be simplified as well.

Change-Id: Id8dfb7b6ee3cc6647e1fa2f14c1fd823b5a6b24b