Add JHC creation when there is no caching set.
The builder now returns a single JsonHomeClient when there is only one
host and a WSClient added.
To make this work, the Builder stores the amount (zero, one or many) of
added hosts in its type. This way the builder fails in at compile-time in
situations that are not valid (more than one host but no caching configured).